backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
scikit-learn==1.7.2
pandas==2.3.3
numpy==2.3.4
requests==2.32.5
python-whois==0.9.6
groq==0.13.0
python-dotenv==1.0.1
gunicorn==23.0.0
